PID, Names, State, nvcsw, nivcsw - Complete
The Children handling needs work, as it is not counting that correctly.

Variable Current, will give all the information of the current process in the form of task_struct. Thus on line 49 I made it equal task.
If we did not do so, current = current->parent does not work as we cannot manipulate current. 

Mostly changed the current variables into task.
current == current->parent does not work since current will never change its value in a process.

Needs work in counting children properly