Made siblings counter start at -1 instead of 0.
So that the current process isn't counted in sibling count. I.E a parent has one child that child has 0 siblings rather than 1, that it had before. Also removed temp and it was redundant. Also as the root doesn't have a sibling list, it doesn't increment to 0, so therefore had to make if statement to check if count_sibling if less than 0 if it is make it 0.